.cursor/skills/digest/requirements.txt
feedparser>=6.0.0
httpx>=0.25.0
pyyaml>=6.0
jinja2>=3.0.0
python-dateutil>=2.8
beautifulsoup4>=4.12
